+++ tests/LegacyDataParserTest.php
@@ -1,25 +1,23 @@
 <?php
 
-require_once 'DataParserTestCase.php';
-
-use PHPUnit\Framework\DataParserTestCase;
+use PHPUnit\Framework\TestCase;
 
 /**
  * Legacy data parser additional tests
  */
-class LegacyDataParserTest extends DataParserTestCase
+class LegacyDataParserTest extends TestCase
 {
     public function testLegacyDataParser()
     {
         $raw = file_get_contents(__DIR__ . '/resources/validLegacyData/SMS.xml');
         $data = \cstuder\ParseHydrodaten\LegacyDataParser::parse($raw);
 
-        $this->assertEquals((504 - 3) * 2, count($data->values));
-        $this->assertEquals(5, count($this->collectParameters($data)));
-        $this->assertEquals(238, count($this->collectLocations($data)));
-        $this->assertEquals(34, count($this->collectTimestamps($data)));
+        $this->assertEquals((504 - 3) * 2, $data->getCount());
+        $this->assertEquals(5, count($data->getParameters()));
+        $this->assertEquals(238, count($data->getLocations()));
+        $this->assertEquals(34, count($data->getTimestamps()));
 
-        $this->assertContainsOnlyInstancesOf('cstuder\ParseValueholder\Value', $data->values);
+        $this->assertContainsOnlyInstancesOf('cstuder\ParseValueholder\Value', $data->getValues());
     }
 
     public function testParseWert()