+++ tests/SuperParserTest.php
@@ -1,24 +1,23 @@
 <?php
-require_once 'DataParserTestCase.php';
 
-use PHPUnit\Framework\DataParserTestCase;
+use PHPUnit\Framework\TestCase;
 
 /**
  * Quantitative tests of the super parser
  */
-class SuperParserTest extends DataParserTestCase
+class SuperParserTest extends TestCase
 {
     public function testSuperParser()
     {
         $raw = file_get_contents(__DIR__ . '/resources/validData/hydroweb.xml');
         $data = \cstuder\ParseHydrodaten\SuperParser::parse($raw);
 
-        $this->assertEquals((504 - 3) * 2, count($data->values));
-        $this->assertEquals(5, count($this->collectParameters($data)));
-        $this->assertEquals(238, count($this->collectLocations($data)));
-        $this->assertEquals(34, count($this->collectTimestamps($data)));
+        $this->assertEquals((504 - 3) * 2, $data->getCount());
+        $this->assertEquals(5, count($data->getParameters()));
+        $this->assertEquals(238, count($data->getLocations()));
+        $this->assertEquals(34, count($data->getTimestamps()));
 
-        $this->assertContainsOnlyInstancesOf('cstuder\ParseValueholder\Value', $data->values);
+        $this->assertContainsOnlyInstancesOf('cstuder\ParseValueholder\Value', $data->getValues());
     }
 
     public function testSuperParserWithPreciseData()
@@ -26,12 +25,12 @@ class SuperParserTest extends DataParserTestCase
         $raw = file_get_contents(__DIR__ . '/resources/validData/hydroweb_prec.xml');
         $data = \cstuder\ParseHydrodaten\SuperParser::parse($raw);
 
-        $this->assertEquals(477, count($data->values));
-        $this->assertEquals(5, count($this->collectParameters($data)));
-        $this->assertEquals(226 - 6, count($this->collectLocations($data))); // 6 locations do not deliver data
-        $this->assertEquals(5, count($this->collectTimestamps($data)));
+        $this->assertEquals(477, $data->getCount());
+        $this->assertEquals(5, count($data->getParameters()));
+        $this->assertEquals(226 - 6, count($data->getLocations())); // 6 locations do not deliver data
+        $this->assertEquals(5, count($data->getTimestamps()));
 
-        $this->assertContainsOnlyInstancesOf('cstuder\ParseValueholder\Value', $data->values);
+        $this->assertContainsOnlyInstancesOf('cstuder\ParseValueholder\Value', $data->getValues());
     }
 
     public function testSuperParserWithNAQUAData()
@@ -39,12 +38,12 @@ class SuperParserTest extends DataParserTestCase
         $raw = file_get_contents(__DIR__ . '/resources/validData/hydroweb.naqua.xml');
         $data = \cstuder\ParseHydrodaten\SuperParser::parse($raw);
 
-        $this->assertEquals(121, count($data->values));
-        $this->assertEquals(6, count($this->collectParameters($data)));
-        $this->assertEquals(79 - 20, count($this->collectLocations($data))); // 20 locations do not deliver data
-        $this->assertEquals(7, count($this->collectTimestamps($data)));
+        $this->assertEquals(121, $data->getCount());
+        $this->assertEquals(6, count($data->getParameters()));
+        $this->assertEquals(79 - 20, count($data->getLocations())); // 20 locations do not deliver data
+        $this->assertEquals(7, count($data->getTimestamps()));
 
-        $this->assertContainsOnlyInstancesOf('cstuder\ParseValueholder\Value', $data->values);
+        $this->assertContainsOnlyInstancesOf('cstuder\ParseValueholder\Value', $data->getValues());
     }
 
     public function testSuperParserWithLegacyData()
@@ -52,11 +51,11 @@ class SuperParserTest extends DataParserTestCase
         $raw = file_get_contents(__DIR__ . '/resources/validLegacyData/SMS.xml');
         $data = \cstuder\ParseHydrodaten\SuperParser::parse($raw);
 
-        $this->assertEquals((504 - 3) * 2, count($data->values));
-        $this->assertEquals(5, count($this->collectParameters($data)));
-        $this->assertEquals(238, count($this->collectLocations($data)));
-        $this->assertEquals(34, count($this->collectTimestamps($data)));
+        $this->assertEquals((504 - 3) * 2, $data->getCount());
+        $this->assertEquals(5, count($data->getParameters()));
+        $this->assertEquals(238, count($data->getLocations()));
+        $this->assertEquals(34, count($data->getTimestamps()));
 
-        $this->assertContainsOnlyInstancesOf('cstuder\ParseValueholder\Value', $data->values);
+        $this->assertContainsOnlyInstancesOf('cstuder\ParseValueholder\Value', $data->getValues());
     }
 }