+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.0] - 2023-02-04
+
+- Updates `cstuder/parse-valueholder` to V0.2.0.
+- Simplifies tests.
+
 ## [2.0.0] - 2023-01-30
 
 - Returns `cstuder/parse-valueholder` values and rows from the data parsers.